--- src/main/java/racingcar/Main.java
@@ -1,11 +1,0 @@
-package racingcar;
-
-public class Main {
-
-	public static void main(final String[] args) {
-		final RacingCarGame racingCarGame = new RacingCarGame();
-		racingCarGame.reset();
-		racingCarGame.doGame();
-		racingCarGame.afterGame();
-	}
-}
