+++ src/main/java/racinggame/Application.java
@@ -3,6 +3,9 @@ package racinggame;
 public class Application {
 
 	public static void main(final String[] args) {
-		// TODO 자동차 경주 게임 구현
+		final RacingCarGame racingCarGame = new RacingCarGame();
+		racingCarGame.reset();
+		racingCarGame.doGame();
+		racingCarGame.afterGame();
 	}
 }
