+++ src/main/java/racinggame/car/Name.java
@@ -1,7 +1,7 @@
-package racingcar.car;
+package racinggame.car;
 
-import racingcar.common.exception.ErrorCode;
-import racingcar.common.exception.ValidationException;
+import racinggame.exception.ErrorCode;
+import racinggame.exception.ValidationException;
 
 public class Name {
