+++ src/main/java/racinggame/input/ParseUtil.java
@@ -1,4 +1,4 @@
-package racingcar.input;
+package racinggame.input;
 
 import java.util.Arrays;
 import java.util.List;
