+++ src/test/java/racingcar/RacingCarGameTest.java
@@ -6,7 +6,8 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import racingcar.car.Cars;
+import racinggame.car.Cars;
+import racinggame.RacingCarGame;
 
 @DisplayName("레이싱카 게임")
 class RacingCarGameTest {
