+++ src/test/java/racingcar/car/CarTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
+import racinggame.car.Car;
+
 @DisplayName("자동차")
 public class CarTest {
