+++ src/test/java/racingcar/car/DistanceTest.java
@@ -6,7 +6,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import racingcar.common.type.DriveType;
+import racinggame.type.DriveType;
+import racinggame.car.Distance;
 
 /**
  * ParameterizedTest 와 Stream<Arguments> 을 사용하면 테스트 코드를 간소화할 수 있지만,
