+++ src/test/java/racingcar/car/EngineTest.java
@@ -10,10 +10,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import racingcar.common.exception.ErrorCode;
-import racingcar.common.exception.ValidationException;
-import racingcar.common.type.DriveType;
-import racingcar.config.Property;
+import racinggame.exception.ErrorCode;
+import racinggame.exception.ValidationException;
+import racinggame.type.DriveType;
+import racinggame.config.Property;
+import racinggame.car.Engine;
 
 @DisplayName("엔진")
 class EngineTest {
